Nectar 0.2: original JAudio engine, 120 FPS option and Windows port
Audio now runs the game's own JAudio engine instead of a reimplementation.
The sequencer, instrument banks, oscillators and channel manager come from
the decompilation unchanged; only the GameCube DSP boundary is replaced by a
software renderer, adapted from pikmin-nextos and driven through SDL2.
Music, sound effects and cinematic audio all play.

Enabled with -DPIKMIN_NATIVE_JAUDIO=ON, which the Linux packaging script now
passes. With the option off the previous native mixer is used and nothing
changes.

Frame rate: a 120 FPS option joins 30 and 60. The game integrates by elapsed
time, so a higher tick rate does not speed up gameplay.

Windows: cross-compiles to a native x86-64 executable with MinGW-w64. Seven
portability problems fixed, among them std::sqrtf outside <cmath> on MinGW,
a wingdi.h stub shadowing the Windows SDK header, OpenGL entry points beyond
1.1 that opengl32 does not export, and clock_gettime.

Launcher: the installer looked for pikmin/pikmin.real while the build has
produced nectar/nectar.real since the rename, so installation failed. The
package wrappers also exported NECTAR_EXECUTABLE_PATH while the launcher read
PIKMIN_EXECUTABLE_PATH, leaving the launcher unable to locate its own package.
Extraction aborted on Windows at the first Shift-JIS filename in the disc's
FST; disc names are now decoded explicitly.

Docs: AI_HANDOFF.md claimed 30 FPS gameplay and contradicted SIGUIENTE_SESION.
It now records how the frame rate actually works, and that the TIME-002
interpolation machinery is built but never invoked.

+++ include/jaudio/dspdriver.h
@@ -3,14 +3,15 @@
 
 #include "jaudio/audiostruct.h"
 #include "types.h"
+#include <stdint.h>
 
 /////////// JAUDIO DSP DRIVER DEFINITIONS ///////////
 // Global functions (all C++, so no extern C wrap).
 
 dspch_* GetDSPchannelHandle(u32 idx);
 void InitDSPchannel();
-dspch_* AllocDSPchannel(u32 channelMode, u32 ownerId);
-int DeAllocDSPchannel(dspch_*, u32);
+dspch_* AllocDSPchannel(u32 channelMode, uintptr_t ownerId);
+int DeAllocDSPchannel(dspch_*, uintptr_t);
 dspch_* GetLowerDSPchannel();
 dspch_* GetLowerActiveDSPchannel();
 BOOL ForceStopDSPchannel(dspch_*);
